% Generated by tinyroxygen: do not edit by hand
% Please edit documentation in cpp4r.R
\name{add_vec_for_}
\alias{add_vec_for_}
\title{Sum Doubles on 'C++' Side}
\usage{
add_vec_for_(x, num)
}

\arguments{
\item{x}{vector of doubles (R)}

\item{num}{single double (C++)}
}

\description{
Test suite
}

