\name{cpp4rtest-internal-functions}
\alias{_cpp4rtest_Rcpp_grow_}
\alias{_cpp4rtest_Rcpp_push_and_truncate_}
\alias{_cpp4rtest_Rcpp_release_}
\alias{_cpp4rtest_Rcpp_sum_dbl_accumulate_}
\alias{_cpp4rtest_Rcpp_sum_dbl_for_}
\alias{_cpp4rtest_Rcpp_sum_dbl_foreach_}
\alias{_cpp4rtest_Rcpp_sum_int_for_}
\alias{_cpp4rtest_assign_Rcpp_}
\alias{_cpp4rtest_assign_cpp4r_}
\alias{_cpp4rtest_cpp4r_add_vec_for_}
\alias{_cpp4rtest_cpp4r_insert_}
\alias{_cpp4rtest_cpp4r_push_and_truncate_}
\alias{_cpp4rtest_cpp4r_release_}
\alias{_cpp4rtest_cpp4r_safe_}
\alias{_cpp4rtest_data_frame_}
\alias{_cpp4rtest_grow_}
\alias{_cpp4rtest_grow_cplx_}
\alias{_cpp4rtest_grow_strings_Rcpp_}
\alias{_cpp4rtest_grow_strings_cpp4r_}
\alias{_cpp4rtest_grow_strings_manual_}
\alias{_cpp4rtest_notroxcpp1_}
\alias{_cpp4rtest_notroxcpp6_}
\alias{_cpp4rtest_ordered_map_to_list_}
\alias{_cpp4rtest_protect_many_}
\alias{_cpp4rtest_protect_many_Rcpp_}
\alias{_cpp4rtest_protect_many_cpp4r_}
\alias{_cpp4rtest_protect_many_preserve_}
\alias{_cpp4rtest_protect_many_sexp_}
\alias{_cpp4rtest_protect_one_}
\alias{_cpp4rtest_protect_one_cpp4r_}
\alias{_cpp4rtest_protect_one_preserve_}
\alias{_cpp4rtest_protect_one_sexp_}
\alias{_cpp4rtest_string_proxy_assignment_}
\alias{_cpp4rtest_string_push_back_}
\alias{_cpp4rtest_sum_cplx_accumulate_}
\alias{_cpp4rtest_sum_cplx_for2_}
\alias{_cpp4rtest_sum_cplx_for_}
\alias{_cpp4rtest_sum_cplx_for_2_}
\alias{_cpp4rtest_sum_cplx_for_3_}
\alias{_cpp4rtest_sum_cplx_for_4_}
\alias{_cpp4rtest_sum_cplx_for_5_}
\alias{_cpp4rtest_sum_cplx_for_6_}
\alias{_cpp4rtest_sum_cplx_foreach_}
\alias{_cpp4rtest_sum_dbl_accumulate2_}
\alias{_cpp4rtest_sum_dbl_accumulate_}
\alias{_cpp4rtest_sum_dbl_for2_}
\alias{_cpp4rtest_sum_dbl_for3_}
\alias{_cpp4rtest_sum_dbl_for_}
\alias{_cpp4rtest_sum_dbl_foreach2_}
\alias{_cpp4rtest_sum_dbl_foreach_}
\alias{_cpp4rtest_sum_int_accumulate_}
\alias{_cpp4rtest_sum_int_for2_}
\alias{_cpp4rtest_sum_int_for_}
\alias{_cpp4rtest_sum_int_foreach_}
\alias{_cpp4rtest_unordered_map_to_list_}
\alias{Rcpp_grow_}
\alias{Rcpp_push_and_truncate_}
\alias{Rcpp_release_}
\alias{Rcpp_sum_dbl_accumulate_}
\alias{Rcpp_sum_dbl_for_}
\alias{Rcpp_sum_dbl_foreach_}
\alias{Rcpp_sum_int_for_}
\alias{assign_Rcpp_}
\alias{assign_cpp4r_}
\alias{cpp4r_add_vec_for_}
\alias{cpp4r_insert_}
\alias{cpp4r_push_and_truncate_}
\alias{cpp4r_release_}
\alias{cpp4r_safe_}
\alias{data_frame_}
\alias{grow_}
\alias{grow_cplx_}
\alias{grow_strings_Rcpp_}
\alias{grow_strings_cpp4r_}
\alias{grow_strings_manual_}
\alias{notroxcpp1_}
\alias{notroxcpp6_}
\alias{ordered_map_to_list_}
\alias{protect_many_}
\alias{protect_many_Rcpp_}
\alias{protect_many_cpp4r_}
\alias{protect_many_preserve_}
\alias{protect_many_sexp_}
\alias{protect_one_}
\alias{protect_one_cpp4r_}
\alias{protect_one_preserve_}
\alias{protect_one_sexp_}
\alias{string_proxy_assignment_}
\alias{string_push_back_}
\alias{sum_cplx_accumulate_}
\alias{sum_cplx_for2_}
\alias{sum_cplx_for_}
\alias{sum_cplx_for_2_}
\alias{sum_cplx_for_3_}
\alias{sum_cplx_for_4_}
\alias{sum_cplx_for_5_}
\alias{sum_cplx_for_6_}
\alias{sum_cplx_foreach_}
\alias{sum_dbl_accumulate2_}
\alias{sum_dbl_accumulate_}
\alias{sum_dbl_for2_}
\alias{sum_dbl_for3_}
\alias{sum_dbl_for_}
\alias{sum_dbl_foreach2_}
\alias{sum_dbl_foreach_}
\alias{sum_int_accumulate_}
\alias{sum_int_for2_}
\alias{sum_int_for_}
\alias{sum_int_foreach_}
\alias{unordered_map_to_list_}

\title{Internal C++ Test Functions}

\description{
These are internal C++ functions used for benchmarking and testing the cpp4r package functionality. They are exported for testing purposes but are not intended for end-user use.
}

\details{
These functions are automatically generated by the cpp4r registration system and are used internally for:
\itemize{
  \item Performance benchmarking between Rcpp and cpp4r implementations
  \item Testing memory protection mechanisms
  \item Testing vector growth and manipulation
  \item Testing string operations
  \item Testing complex number operations
}

Users should not call these functions directly. They are documented here only to satisfy R CMD check requirements.
}

\keyword{internal}
