% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpp4rtest-package.R
\docType{package}
\name{cpp4rtest-package}
\alias{cpp4rtest}
\alias{cpp4rtest-package}
\title{cpp4rtest: Testing 'cpp4r'}
\description{
To check cpp4r correctness. The headers are vendored to check which lines of the C++ code are covered by the unit tests.
}
\author{
\strong{Maintainer}: Mauricio Vargas Sepulveda \email{m.vargas.sepulveda@gmail.com} (\href{https://orcid.org/0000-0003-1017-7574}{ORCID})

}
\keyword{internal}
