% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpp4r.R
\name{env_get_int_}
\alias{env_get_int_}
\title{Get Integer from Environment on 'C++' Side}
\usage{
env_get_int_(env, name)
}
\arguments{
\item{env}{R environment to query}

\item{name}{name of the variable to get}
}
\description{
Test suite
}
