% Generated by tinyroxygen: do not edit by hand
% Please edit documentation in cpp4r.R
\name{env_get_str_}
\alias{env_get_str_}
\title{Get String from Environment on 'C++' Side}
\usage{
env_get_str_(env, name)
}

\arguments{
\item{env}{R environment to query}

\item{name}{name of the variable to get}
}

\description{
Test suite
}

