% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpp4r.R
\name{find_name_pos_}
\alias{find_name_pos_}
\title{Find Position of Name in Named Vector on 'C++' Side}
\usage{
find_name_pos_(x, name)
}
\arguments{
\item{x}{named vector to query}

\item{name}{name of the element to find}
}
\description{
Test suite
}
