% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpp4r.R
\name{findInterval2}
\alias{findInterval2}
\title{Find Interval for Multiple Values on 'C++' Side}
\usage{
findInterval2(x, breaks)
}
\arguments{
\item{x}{vector of values to find intervals for}

\item{breaks}{vector of break points}
}
\description{
Test suite
}
