% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpp4r.R
\name{findInterval3}
\alias{findInterval3}
\title{Find Interval for Multiple Values on 'C++' Side (optimized, raw pointers)}
\usage{
findInterval3(x, breaks)
}
\arguments{
\item{x}{vector of values to find intervals for}

\item{breaks}{vector of break points}
}
\description{
Test suite
}
