% Generated by tinyroxygen: do not edit by hand
% Please edit documentation in cpp4r.R
\name{get_by_name_}
\alias{get_by_name_}
\title{Get Value by Name from Named Vector on 'C++' Side}
\usage{
get_by_name_(x, name)
}

\arguments{
\item{x}{named vector to query}

\item{name}{name of the element to get}
}

\description{
Test suite
}

