% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpp4r.R
\name{grow_strings_}
\alias{grow_strings_}
\title{Grow Strings}
\usage{
grow_strings_(n, seed)
}
\arguments{
\item{n}{number of strings to grow}

\item{seed}{seed for random string generation}
}
\description{
Test suite
}
