% Generated by tinyroxygen: do not edit by hand
% Please edit documentation in cpp4r.R
\name{iterator_at_}
\alias{iterator_at_}
\title{Access Element in Double Vector}
\usage{
iterator_at_(x, index)
}

\arguments{
\item{x}{vector of double numbers (R)}

\item{index}{index of the element to access (1-based)}
}

\description{
Test suite
}

