% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpp4r.R
\name{iterator_double_values_}
\alias{iterator_double_values_}
\title{Double Values in Double Vector}
\usage{
iterator_double_values_(x)
}
\arguments{
\item{x}{vector of double numbers (R)}
}
\description{
Test suite
}
