% Generated by tinyroxygen: do not edit by hand
% Please edit documentation in cpp4r.R
\name{my_stop_n2_}
\alias{my_stop_n2_}
\title{Errors with Two Arguments}
\usage{
my_stop_n2_(mystring, myarg)
}

\arguments{
\item{mystring}{string to include in the error}

\item{myarg}{additional string to include in the error}
}

\description{
Test suite
}

