% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpp4r.R
\name{protect_many_preserve_}
\alias{protect_many_preserve_}
\title{Protect Many Objects Using R_PreserveObject}
\usage{
protect_many_preserve_(n)
}
\arguments{
\item{n}{number of objects to protect}
}
\description{
Test suite
}
