% Generated by tinyroxygen: do not edit by hand
% Please edit documentation in cpp4r.R
\name{protect_many_sexp_}
\alias{protect_many_sexp_}
\title{Protect Many Objects Using cpp4r SEXP}
\usage{
protect_many_sexp_(n)
}

\arguments{
\item{n}{number of objects to protect}
}

\description{
Test suite
}

