% Generated by tinyroxygen: do not edit by hand
% Please edit documentation in cpp4r.R
\name{protect_one_preserve_}
\alias{protect_one_preserve_}
\title{Protect One Object Using R_PreserveObject}
\usage{
protect_one_preserve_(x, n)
}

\arguments{
\item{x}{object to protect}

\item{n}{number of times to protect}
}

\description{
Test suite
}

