% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpp4r.R
\name{roxcpp2}
\alias{roxcpp2}
\title{Roxygenise C++ function II}
\usage{
roxcpp2(x)
}
\arguments{
\item{x}{numeric value}
}
\description{
Dummy function to test roxygen2. It adds 2.0 to a double.
}
\examples{
roxcpp2(1.0)
}
