% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpp4r.R
\name{roxcpp3}
\alias{roxcpp3}
\title{Roxygenise C++ function III}
\usage{
roxcpp3(x)
}
\arguments{
\item{x}{numeric value}
}
\description{
Dummy function to test roxygen2. It adds 3.0 to a double.
}
\examples{
roxcpp3(1.0)
}
