% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpp4r.R
\name{roxcpp4}
\alias{roxcpp4}
\title{Roxygenise C++ function IV}
\usage{
roxcpp4(x)
}
\arguments{
\item{x}{numeric value}
}
\description{
Dummy function to test roxygen2. It adds 4.0 to a double.
}
\examples{
roxcpp4(1.0)
}
