% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpp4r.R
\name{roxcpp5}
\alias{roxcpp5}
\title{Roxygenise C++ function V}
\usage{
roxcpp5(x)
}
\arguments{
\item{x}{numeric value}
}
\description{
Dummy function to test roxygen2. It adds 5.0 to a double.
}
\examples{
roxcpp5(1.0)
}
