% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpp4r.R
\name{roxcpp7}
\alias{roxcpp7}
\title{Roxygenise C++ function VII}
\usage{
roxcpp7(x)
}
\arguments{
\item{x}{numeric value}
}
\description{
Dummy function to test roxygen2. It adds 7.0 to a double.
}
\examples{
my_fun <- function(x) {
  roxcpp7(x)
}
}
\seealso{
\code{\link{roxcpp3}}
}
