% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpp4r.R
\name{sum_cplx_complexes_out_}
\alias{sum_cplx_complexes_out_}
\title{Sum Complex Numbers on 'C++' Side (typed in, complexes out)}
\usage{
sum_cplx_complexes_out_(x)
}
\arguments{
\item{x}{vector of complex numbers (R)}
}
\description{
Test suite
}
