% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpp4r.R
\name{sum_cplx_foreach_}
\alias{sum_cplx_foreach_}
\title{Sum Complex Numbers on 'C++' Side (foreach, std out)}
\usage{
sum_cplx_foreach_(x)
}
\arguments{
\item{x}{vector of complex numbers (R)}
}
\description{
Test suite
}
