% Generated by tinyroxygen: do not edit by hand
% Please edit documentation in cpp4r.R
\name{sum_cplx_r_complex_out_}
\alias{sum_cplx_r_complex_out_}
\title{Sum Complex Numbers on 'C++' Side (typed in, r_complex out)}
\usage{
sum_cplx_r_complex_out_(x)
}

\arguments{
\item{x}{vector of complex numbers (R)}
}

\description{
Test suite
}

