% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpp4r.R
\name{sum_cplx_typed_std_out_}
\alias{sum_cplx_typed_std_out_}
\title{Sum Complex Numbers on 'C++' Side (typed in re-wrapped, std out)}
\usage{
sum_cplx_typed_std_out_(x_sxp)
}
\arguments{
\item{x_sxp}{vector of complex numbers (R)}
}
\description{
Test suite
}
