% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpp4r.R
\name{sum_dbl_accumulate_}
\alias{sum_dbl_accumulate_}
\title{Sum Double Numbers on 'C++' Side (typed in, double out)}
\usage{
sum_dbl_accumulate_(x)
}
\arguments{
\item{x}{vector of double numbers (R)}
}
\description{
Test suite
}
