% Generated by tinyroxygen: do not edit by hand
% Please edit documentation in cpp4r.R
\name{sum_dbl_for_}
\alias{sum_dbl_for_}
\title{Sum Double Numbers on 'C++' Side (typed in, double out)}
\usage{
sum_dbl_for_(x)
}

\arguments{
\item{x}{vector of double numbers (R)}
}

\description{
Test suite
}

