% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpp4r.R
\name{sum_int_accumulate_}
\alias{sum_int_accumulate_}
\title{Sum Integer Numbers on 'C++' Side (typed in, double out, accumulate)}
\usage{
sum_int_accumulate_(x)
}
\arguments{
\item{x}{vector of integer numbers (R)}
}
\description{
Test suite
}
