% Generated by tinyroxygen: do not edit by hand
% Please edit documentation in cpp4r.R
\name{sum_int_foreach_}
\alias{sum_int_foreach_}
\title{Sum Integer Numbers on 'C++' Side (typed in, double out, foreach)}
\usage{
sum_int_foreach_(x)
}

\arguments{
\item{x}{vector of integer numbers (R)}
}

\description{
Test suite
}

