% Generated by tinyroxygen: do not edit by hand
% Please edit documentation in cpp4r.R
\name{sum_int_sexp_for_}
\alias{sum_int_sexp_for_}
\title{Sum Integer Numbers on 'C++' Side (SEXP in no protect, double out, for loop)}
\usage{
sum_int_sexp_for_(x_sxp)
}

\arguments{
\item{x_sxp}{vector of integer numbers (R)}
}

\description{
Test suite
}

