% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpp4r.R
\name{grow_strings_cpp4r_}
\alias{grow_strings_cpp4r_}
\alias{grow_strings_manual_}
\alias{assign_cpp4r_}
\title{Grow functions}
\usage{
grow_strings_cpp4r_(n, seed)

grow_strings_manual_(n, seed)

assign_cpp4r_(n, seed)
}
\description{
Grow functions

Grow functions

Grow functions
}
\keyword{internal}
