% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpp4r.R
\name{cpp4r_add_int_vec_}
\alias{cpp4r_add_int_vec_}
\alias{cpp4r_as_integers_}
\alias{cpp4r_negate_logical_}
\alias{cpp4r_logical_to_int_}
\alias{cpp4r_logical_to_dbl_}
\alias{cpp4r_iterator_sum_}
\alias{cpp4r_iterator_sum_int_}
\alias{cpp4r_iterator_find_}
\alias{cpp4r_iterator_count_}
\alias{cpp4r_iterator_min_}
\alias{cpp4r_iterator_max_}
\alias{cpp4r_iterator_double_values_}
\alias{cpp4r_reverse_vector_}
\alias{cpp4r_iterator_at_}
\alias{cpp4r_iterator_distance_}
\alias{cpp4r_raw_copy_}
\alias{cpp4r_raw_xor_}
\title{Test helper functions}
\usage{
cpp4r_add_int_vec_(x, value)

cpp4r_as_integers_(x)

cpp4r_negate_logical_(x)

cpp4r_logical_to_int_(x)

cpp4r_logical_to_dbl_(x)

cpp4r_iterator_sum_(x)

cpp4r_iterator_sum_int_(x)

cpp4r_iterator_find_(x, value)

cpp4r_iterator_count_(x, value)

cpp4r_iterator_min_(x)

cpp4r_iterator_max_(x)

cpp4r_iterator_double_values_(x)

cpp4r_reverse_vector_(x)

cpp4r_iterator_at_(x, index)

cpp4r_iterator_distance_(x)

cpp4r_raw_copy_(x)

cpp4r_raw_xor_(x, mask)
}
\description{
Test helper functions

Test helper functions

Test helper functions

Test helper functions

Test helper functions

Test helper functions

Test helper functions

Test helper functions

Test helper functions

Test helper functions

Test helper functions

Test helper functions

Test helper functions

Test helper functions

Test helper functions

Test helper functions

Test helper functions
}
\keyword{internal}
