% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpp4r.R
\name{cpp4r_list_of_doubles_}
\alias{cpp4r_list_of_doubles_}
\alias{cpp4r_list_of_integers_}
\alias{cpp4r_list_of_strings_}
\alias{cpp4r_list_of_named_}
\alias{cpp4r_make_complex_}
\alias{cpp4r_complex_real_}
\alias{cpp4r_complex_imag_}
\alias{cpp4r_complex_modulus_}
\alias{cpp4r_complex_add_}
\title{List and complex helper functions}
\usage{
cpp4r_list_of_doubles_()

cpp4r_list_of_integers_()

cpp4r_list_of_strings_()

cpp4r_list_of_named_()

cpp4r_make_complex_(real, imag)

cpp4r_complex_real_(x)

cpp4r_complex_imag_(x)

cpp4r_complex_modulus_(x)

cpp4r_complex_add_(x, y)
}
\description{
List and complex helper functions

List and complex helper functions

List and complex helper functions

List and complex helper functions

List and complex helper functions

List and complex helper functions

List and complex helper functions

List and complex helper functions

List and complex helper functions
}
\keyword{internal}
