% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpp4r.R
\name{cpp4r_named_list_push_back_}
\alias{cpp4r_named_list_push_back_}
\alias{cpp4r_named_list_c_style_}
\title{List functions}
\usage{
cpp4r_named_list_push_back_()

cpp4r_named_list_c_style_()
}
\description{
List functions

List functions
}
\keyword{internal}
