% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpp4r.R
\name{ordered_map_to_list_}
\alias{ordered_map_to_list_}
\alias{unordered_map_to_list_}
\title{List functions for cpp4r}
\usage{
ordered_map_to_list_(x)

unordered_map_to_list_(x)
}
\description{
List functions for cpp4r

List functions for cpp4r
}
\keyword{internal}
