% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpp4r.R
\name{protect_one_}
\alias{protect_one_}
\alias{protect_one_sexp_}
\alias{protect_one_cpp4r_}
\alias{protect_one_preserve_}
\alias{protect_many_}
\alias{protect_many_cpp4r_}
\alias{protect_many_sexp_}
\alias{protect_many_preserve_}
\title{Protect functions}
\usage{
protect_one_(x, n)

protect_one_sexp_(x, n)

protect_one_cpp4r_(x, n)

protect_one_preserve_(x, n)

protect_many_(n)

protect_many_cpp4r_(n)

protect_many_sexp_(n)

protect_many_preserve_(n)
}
\description{
Protect functions

Protect functions

Protect functions

Protect functions

Protect functions

Protect functions

Protect functions

Protect functions
}
\keyword{internal}
