% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpp4r.R
\name{unordered_map_to_list_}
\alias{unordered_map_to_list_}
\title{Convert Unordered Map to List}
\usage{
unordered_map_to_list_(x)
}
\arguments{
\item{x}{vector of double numbers (R)}
}
\description{
Test suite
}
