% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpp4r.R
\name{add_int_vec_}
\alias{add_int_vec_}
\title{Add Integer Value to Integer Vector}
\usage{
add_int_vec_(x, value)
}
\arguments{
\item{x}{vector of integer numbers (R)}

\item{value}{integer value to add}
}
\description{
Test suite
}
