% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpp4r.R
\name{complex_add_}
\alias{complex_add_}
\title{Add Complex Vectors on 'C++' Side (typed in, complexes out)}
\usage{
complex_add_(x, y)
}
\arguments{
\item{x}{vector of complex numbers}

\item{y}{vector of complex numbers}
}
\description{
Test suite
}
