% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpp4r.R
\name{complex_imag_}
\alias{complex_imag_}
\title{Extract Imaginary Part of Complex Numbers on 'C++' Side (typed in, r_complex out)}
\usage{
complex_imag_(x)
}
\arguments{
\item{x}{vector of complex numbers}
}
\description{
Test suite
}
