% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpp4r.R
\name{complex_modulus_}
\alias{complex_modulus_}
\title{Compute Modulus of Complex Numbers on 'C++' Side (typed in, r_complex out)}
\usage{
complex_modulus_(x)
}
\arguments{
\item{x}{vector of complex numbers}
}
\description{
Test suite
}
