% Generated by tinyroxygen: do not edit by hand
% Please edit documentation in cpp4r.R
\name{complex_real_}
\alias{complex_real_}
\title{Extract Real Part of Complex Numbers on 'C++' Side (typed in, r_complex out)}
\usage{
complex_real_(x)
}

\arguments{
\item{x}{vector of complex numbers}
}

\description{
Test suite
}

