% Generated by tinyroxygen: do not edit by hand
% Please edit documentation in cpp4r.R
\name{contains_name_}
\alias{contains_name_}
\title{Check if Named Vector Contains Name on 'C++' Side}
\usage{
contains_name_(x, name)
}

\arguments{
\item{x}{named vector to query}

\item{name}{name of the element to check}
}

\description{
Test suite
}

