% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpp4rtest-package.R
\docType{package}
\name{cpp4rtest-package}
\alias{cpp4rtest}
\alias{cpp4rtest-package}
\title{cpp4rtest: A test suite and benchmark code for 'cpp4r'}
\description{
Provides a test suite and benchmarking code for the 'cpp4r' package.
}
\author{
\strong{Maintainer}: Jim Hester \email{james.f.hester@gmail.com} (\href{https://orcid.org/0000-0002-2739-7082}{ORCID})

Other contributors:
\itemize{
  \item RStudio [copyright holder, funder]
}

}
\keyword{internal}
