% Generated by tinyroxygen: do not edit by hand
% Please edit documentation in cpp4rtest-package.R
\name{cpp4rtest-package}
\alias{cpp4rtest-package}
\title{Testing 'cpp4r'}
\description{
To check cpp4r correctness. The headers are vendored to check
 which lines ofthe C++ code are covered by the unit tests.
}

