% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpp4rtest-package.R
\docType{package}
\name{cpp4rtest-package}
\alias{cpp4rtest}
\alias{cpp4rtest-package}
\title{cpp4rtest: A Test Suite and Benchmark Code for 'cpp4r'}
\description{
Provides a test suite and benchmarking code for the 'cpp4r' package.
}
\author{
\strong{Maintainer}: Mauricio Vargas Sepulveda \email{m.vargas.sepulveda@gmail.com} (\href{https://orcid.org/0000-0003-1017-7574}{ORCID})

}
\keyword{internal}
