% Generated by tinyroxygen: do not edit by hand
% Please edit documentation in cpp4r.R
\name{env_exists_}
\alias{env_exists_}
\title{Check if Variable Exists in Environment on 'C++' Side}
\usage{
env_exists_(env, name)
}

\arguments{
\item{env}{R environment to query}

\item{name}{name of the variable to check}
}

\description{
Test suite
}

