% Generated by tinyroxygen: do not edit by hand
% Please edit documentation in cpp4r.R
\name{env_set_}
\alias{env_set_}
\title{Set Integer in Environment on 'C++' Side}
\usage{
env_set_(env, name, value)
}

\arguments{
\item{env}{R environment to modify}

\item{name}{name of the variable to set}

\item{value}{integer value to set}
}

\description{
Test suite
}

