% Generated by tinyroxygen: do not edit by hand
% Please edit documentation in cpp4r.R
\name{findInterval2_5}
\alias{findInterval2_5}
\title{Find Interval for Multiple Values on 'C++' Side (optimized)}
\usage{
findInterval2_5(x, breaks)
}

\arguments{
\item{x}{vector of values to find intervals for}

\item{breaks}{vector of break points}
}

\description{
Test suite
}

