% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpp4r.R
\name{gibbs_cpp2_}
\alias{gibbs_cpp2_}
\title{Gibbs Sampler in C++ (Alternative)}
\usage{
gibbs_cpp2_(N, thin)
}
\arguments{
\item{N}{number of iterations}

\item{thin}{thinning parameter}
}
\description{
Test suite
}
