% Generated by tinyroxygen: do not edit by hand
% Please edit documentation in cpp4r.R
\name{iterator_count_}
\alias{iterator_count_}
\title{Count Value in Double Vector}
\usage{
iterator_count_(x, value)
}

\arguments{
\item{x}{vector of double numbers (R)}

\item{value}{double value to count}
}

\description{
Test suite
}

