% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpp4r.R
\name{iterator_find_}
\alias{iterator_find_}
\title{Find Value in Double Vector}
\usage{
iterator_find_(x, value)
}
\arguments{
\item{x}{vector of double numbers (R)}

\item{value}{double value to find}
}
\description{
Test suite
}
