% Generated by tinyroxygen: do not edit by hand
% Please edit documentation in cpp4r.R
\name{make_complex_}
\alias{make_complex_}
\title{Create Complex Numbers on 'C++' Side (typed in, r_complex out)}
\usage{
make_complex_(real, imag)
}

\arguments{
\item{real}{vector of real parts}

\item{imag}{vector of imaginary parts}
}

\description{
Test suite
}

