% Generated by tinyroxygen: do not edit by hand
% Please edit documentation in cpp4r.R
\name{matrix_add_coerce_test_}
\alias{matrix_add_coerce_test_}
\title{Add Two Matrices (with Integer Coercion)}
\usage{
matrix_add_coerce_test_(x, y)
}

\arguments{
\item{x}{first matrix (R)}

\item{y}{second matrix (R)}
}

\description{
Test suite
}

