% Generated by tinyroxygen: do not edit by hand
% Please edit documentation in cpp4r.R
\name{matrix_mixed_add_}
\alias{matrix_mixed_add_}
\title{Add Integer Matrix and Double Matrix}
\usage{
matrix_mixed_add_(int_mat, dbl_mat)
}

\arguments{
\item{int_mat}{integer matrix (R)}

\item{dbl_mat}{double matrix (R)}
}

\description{
Test suite
}

