% Generated by tinyroxygen: do not edit by hand
% Please edit documentation in cpp4r.R
\name{my_message_n2_}
\alias{my_message_n2_}
\title{Messages with Two Arguments}
\usage{
my_message_n2_(mystring, myarg)
}

\arguments{
\item{mystring}{string to include in the message}

\item{myarg}{additional string to include in the message}
}

\description{
Test suite
}

