% Generated by tinyroxygen: do not edit by hand
% Please edit documentation in cpp4r.R
\name{my_warning_n2_}
\alias{my_warning_n2_}
\title{Warnings with Two Arguments}
\usage{
my_warning_n2_(mystring, myarg)
}

\arguments{
\item{mystring}{string to include in the warning}

\item{myarg}{additional string to include in the warning}
}

\description{
Test suite
}

