% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpp4r.R
\name{push_and_truncate_}
\alias{push_and_truncate_}
\title{Push and Truncate Doubles}
\usage{
push_and_truncate_(size_sexp)
}
\arguments{
\item{size_sexp}{initial size of the vector to grow and truncate}
}
\description{
Test suite
}
