% Generated by tinyroxygen: do not edit by hand
% Please edit documentation in cpp4r.R
\name{sum_cplx_accumulate_}
\alias{sum_cplx_accumulate_}
\title{Sum Complex Numbers on 'C++' Side (accumulate, std out)}
\usage{
sum_cplx_accumulate_(x)
}

\arguments{
\item{x}{vector of complex numbers (R)}
}

\description{
Test suite
}

