% Generated by tinyroxygen: do not edit by hand
% Please edit documentation in cpp4r.R
\name{sum_dbl_sexp_writable_for_}
\alias{sum_dbl_sexp_writable_for_}
\title{Sum Double Numbers on 'C++' Side (SEXP in no protect, double out)}
\usage{
sum_dbl_sexp_writable_for_(x_sxp)
}

\arguments{
\item{x_sxp}{vector of double numbers (R)}
}

\description{
Test suite
}

