% Generated by tinyroxygen: do not edit by hand
% Please edit documentation in cpp4r.R
\name{upper_bound}
\alias{upper_bound}
\title{Find Interval for a Single Value on 'C++' Side (typed in, double out)}
\usage{
upper_bound(x, breaks)
}

\arguments{
\item{x}{single value to find interval for}

\item{breaks}{vector of break points}
}

\description{
Test suite
}

