% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpp4r.R
\name{weak_ref_typeof_}
\alias{weak_ref_typeof_}
\title{Weak Reference Type}
\usage{
weak_ref_typeof_(key, val)
}
\arguments{
\item{key}{key for weak reference}

\item{val}{value for weak reference}
}
\description{
Test suite
}
