% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpp4r.R
\name{weak_ref_value_}
\alias{weak_ref_value_}
\title{Weak Reference Value}
\usage{
weak_ref_value_(key, val)
}
\arguments{
\item{key}{key for weak reference}

\item{val}{value for weak reference}
}
\description{
Test suite
}
